fix(aggregate,scalar)!: compute f32 embeddings in f64 to end the cancellation defect class
Three consecutive review rounds found the same numerical class inside the
previous round's fix, because the contract, not the arithmetic, was wrong: an
f32 embedding accumulated in f32, so every power-of-two scaling scheme had a
next adversarial counterexample (subnormal flush, exact-cancellation
fabrication, mixed-magnitude EMA collapse).

Set `f32`'s associated compute type to `f64`. `f32` becomes a storage-only
scalar and is no longer a `Real`; `f64` is the sole compute domain, and both
shipped scalars (and the int8 test double) compute there. Every f32 is exact in
f64, every f32 subnormal is a normal f64, and f32::MAX squared sits far inside
f64's range, so the entire per-fold power-of-two shift apparatus becomes
unnecessary and is deleted: accumulation_shift, ema_shift, saliency_shift,
scaled_l2_norm, and Real's MAX_EXP/MIN_EXP all go.

Two hazards survive into f64 itself and are handled where they arise:

- Cancellation across a wide exponent spread now uses compensated (Neumaier)
  summation, so an exactly cancelling weighted sum lands at exactly zero under
  every association order and NonFinite keeps meaning "no direction".
- A norm unrepresentable although the vector is finite is still handled by the
  scale-aware renormalization, which never forms that norm.

The one irreducible ceiling is SaliencyWeighted, which squares magnitudes: its
f64 window is sqrt(MIN_POSITIVE)..sqrt(MAX) (~1.5e-162..~1.3e154), outside which
it returns NonFinite rather than rescale with the fabrication-prone shift that
widening to f64 exists to retire. Every realistic magnitude is far inside it.

The policy trait's compute-scalar parameter now defaults to f64 (the domain
every ordinary embedding computes in); the bare `dyn AggregatePolicy` spelling
and object safety are unchanged. AggregatePolicyKind and its serde wire format
are untouched — configuration stays f32 and widens at use.

Also sweep every growing collection on a Result-returning path in aggregate for
fallibility (Codex R4 MEDIUM): the coverage/borrow/widen buffers in `aggregate`,
the accumulator and compensation in weighted_sum_renorm, EMA's state, and
saliency's weights now grow through try_reserve and map a refused allocation to
WinditError::AllocFailed. A new tests/aggregate_alloc.rs arms a refusing global
allocator and proves each built-in policy reports AllocFailed rather than
aborting.

The f32 bit-exactness guard, which pinned bit-identity with naive f32 folding,
is replaced: accumulating in f64 and rounding once is deliberately more accurate
than an f32 fold, so the guard now asserts the scale-aware result is within one
ulp of the naive f64 fold and L2-unit to f64 precision.

+++ src/windowed/tests.rs
@@ -45,11 +45,11 @@ fn zero_norm_is_nonfinite_error() {
 #[test]
 fn non_finite_input_errors() {
   assert!(matches!(
-    TestVec::from_unnormalized(&[f32::INFINITY, 1.0]),
+    TestVec::from_unnormalized(&[f64::INFINITY, 1.0]),
     Err(WinditError::NonFinite)
   ));
   assert!(matches!(
-    TestVec::from_unnormalized(&[f32::NAN, 1.0]),
+    TestVec::from_unnormalized(&[f64::NAN, 1.0]),
     Err(WinditError::NonFinite)
   ));
 }
